Various edits and fixes
