--- mod/bigworld/BigWorld.py
@@ -1,16 +1,0 @@
-class BigWorld():
-    pass
-
-def player():
-    return None
-
-
-entities = {}
-
-
-def serverTime():
-    return None
-
-
-def LatencyInfo():
-    return None
